refactor: Line 이 일급 컬렉션을 사용하도록 리팩터링
- Sections : Section 을 컬렉션의 형태로 가지는 일급컬렉션
  - Line 에서 Section 에 관련된 로직을 이동시킴
- Line : Section 관련 로직을 Sections 로 이동
- 테스트들에 노출된 Section 관련된 테스트를 Station 으로 변경하여 외부에서 알 필요가 없는 Section 에 대한 노출을 줄임
